# packages/shared/schema/openapi.json
{
  "components": {
    "schemas": {
...
      "ConsentResponse": {
        "properties": {
          "consent_version": {
            "title": "Consent Version",
            "type": "string"
          },
          "data_processing": {
            "title": "Data Processing",
            "type": "boolean"
          },
          "granted": {
            "title": "Granted",
            "type": "boolean"
          },
          "photos": {
            "title": "Photos",
            "type": "boolean"
          },
          "privacy": {
            "title": "Privacy",
            "type": "boolean"
          },
          "revoked": {
            "title": "Revoked",
            "type": "boolean"
          }
        },
        "required": [
          "granted",
          "privacy",
          "photos",
          "data_processing",
          "consent_version",
          "revoked"
        ],
        "title": "ConsentResponse",
        "type": "object"
      },
      "CreateBookResponse": {
        "additionalProperties": false,
        "properties": {
          "estimated_time_seconds": {
            "anyOf": [
              {
                "maximum": 3600.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Time Seconds"
          },
          "job_id": {
            "maxLength": 60,
            "minLength": 1,
            "title": "Job Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/JobState"
          }
        },
        "required": [
          "job_id",
          "status"
        ],
        "title": "CreateBookResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JobState": {
        "enum": [
          "queued",
          "running",
          "failed",
          "done"
        ],
        "title": "JobState",
        "type": "string"
      },
      "JobStatus": {
        "additionalProperties": false,
        "properties": {
          "current_step": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Current Step",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorInfo"
              },
              {
                "type": "null"
              }
            ]
          },
          "job_id": {
            "maxLength": 60,
            "minLength": 1,
            "title": "Job Id",
            "type": "string"
          },
          "progress": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Progress",
            "type": "integer"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BookResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "$ref": "#/components/schemas/JobState"
          }
        },
        "required": [
          "job_id",
          "status",
          "progress",
          "current_step"
        ],
        "title": "JobStatus",
        "type": "object"
      },
...
    }
  },
  "info": {
    "contact": {
      "email": "support@aistorybook.com",
      "name": "AI Story Book Team"
    },
    "description": "\n# AI Story Book API\n\nAI 기반 맞춤형 동화책 생성 API입니다.\n\n## 주요 기능\n\n* **책 생성**: 주제, 연령대, 스타일을 입력하면 AI가 동화책을 생성합니다\n* **캐릭터 관리**: 사진에서 캐릭터를 추출하거나 직접 생성할 수 있습니다\n* **시리즈 생성**: 같은 캐릭터로 연속된 이야기를 만들 수 있습니다\n* **PDF/오디오 내보내기**: 완성된 책을 PDF나 오디오로 내보낼 수 있습니다\n\n## 인증\n\n모든 API는 `X-User-Key` 헤더가 필요합니다.\n중복 요청 방지를 위해 `X-Idempotency-Key` 헤더 사용을 권장합니다.\n\n## Rate Limiting\n\n- 기본: 분당 10회 요청 제한\n- 초과 시 429 Too Many Requests 응답\n\n## 크레딧 시스템\n\n책 1권 생성에 크레딧 1개가 소모됩니다.\n    ",
    "license": {
      "name": "MIT"
    },
    "title": "AI Story Book API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/books": {
      "post": {
        "description": "새 동화책 생성 요청\n\n- 비동기로 처리되며 job_id 반환\n- GET /v1/books/{job_id}로 상태 조회\n- 크레딧 1개 필요",
        "operationId": "create_book_v1_books_post",
        "parameters": [
          {
            "description": "User identification key",
            "in": "header",
            "name": "x-user-key",
            "required": true,
            "schema": {
              "description": "User identification key",
              "title": "X-User-Key",
              "type": "string"
            }
          },
          {
            "description": "Optional active child profile id",
            "in": "header",
            "name": "x-profile-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional active child profile id",
              "title": "X-Profile-Id"
            }
          },
          {
            "in": "header",
            "name": "x-idempotency-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookSpec"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateBookResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Book",
        "tags": [
          "Books"
        ]
      }
    },
...
    "/v1/books/{job_id}": {
      "get": {
        "description": "책 생성 상태 조회\n\n- status: queued, running, failed, done\n- progress: 0-100\n- done일 경우 result에 BookResult 포함",
        "operationId": "get_book_status_v1_books__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "description": "User identification key",
            "in": "header",
            "name": "x-user-key",
            "required": true,
            "schema": {
              "description": "User identification key",
              "title": "X-User-Key",
              "type": "string"
            }
          },
          {
            "description": "Optional active child profile id",
            "in": "header",
            "name": "x-profile-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional active child profile id",
              "title": "X-Profile-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Book Status",
        "tags": [
          "Books"
        ]
      }
    },
...
    "/v1/consent": {
      "get": {
        "description": "현재 유효한 보호자 동의 상태 조회.",
        "operationId": "get_consent_v1_consent_get",
        "parameters": [
          {
            "description": "User identification key",
            "in": "header",
            "name": "x-user-key",
            "required": true,
            "schema": {
              "description": "User identification key",
              "title": "X-User-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Consent",
        "tags": [
          "Consent"
        ]
      },
      "post": {
        "description": "보호자 동의 기록. granted 는 필수 동의(개인정보+데이터 처리) 충족 시 True.\n\n유저당 비철회 동의 행은 최대 1개가 되도록, 새 동의 시 기존 비철회 행을 폐기(supersede)한다.",
        "operationId": "grant_consent_v1_consent_post",
        "parameters": [
          {
            "description": "User identification key",
            "in": "header",
            "name": "x-user-key",
            "required": true,
            "schema": {
              "description": "User identification key",
              "title": "X-User-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsentGrantRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Grant Consent",
        "tags": [
          "Consent"
        ]
      }
    },
    "/v1/consent/revoke": {
      "post": {
        "description": "보호자 동의 철회 + 수집된 아동 사진/그림 파생 캐릭터 파기(PIPA/COPPA 철회 의무).\n\n이후 사진 기반 기능은 다시 차단되고, from_photo 캐릭터 레코드·원본 사진(스토리지)은 삭제된다.",
        "operationId": "revoke_consent_v1_consent_revoke_post",
        "parameters": [
          {
            "description": "User identification key",
            "in": "header",
            "name": "x-user-key",
            "required": true,
            "schema": {
              "description": "User identification key",
              "title": "X-User-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revoke Consent",
        "tags": [
          "Consent"
        ]
      }
    },
...
    "/v1/growth": {
      "get": {
        "description": "아이의 읽기 성장 리포트 — 읽은 책·스트릭·학습 어휘·퀴즈 정확도·추정 읽기레벨.",
        "operationId": "get_growth_v1_growth_get",
        "parameters": [
          {
            "description": "User identification key",
            "in": "header",
            "name": "x-user-key",
            "required": true,
            "schema": {
              "description": "User identification key",
              "title": "X-User-Key",
              "type": "string"
            }
          },
          {
            "description": "Optional active child profile id",
            "in": "header",
            "name": "x-profile-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional active child profile id",
              "title": "X-Profile-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Growth",
        "tags": [
          "Growth"
        ]
      }
    },
    "/v1/growth/answers": {
      "post": {
        "description": "학습 퀴즈/어휘 응답 기록 — 성장 측정의 근거 데이터.",
        "operationId": "record_answer_v1_growth_answers_post",
        "parameters": [
          {
            "description": "User identification key",
            "in": "header",
            "name": "x-user-key",
            "required": true,
            "schema": {
              "description": "User identification key",
              "title": "X-User-Key",
              "type": "string"
            }
          },
          {
            "description": "Optional active child profile id",
            "in": "header",
            "name": "x-profile-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional active child profile id",
              "title": "X-Profile-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnswerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Record Answer",
        "tags": [
          "Growth"
        ]
      }
    },
    "/v1/growth/peers": {
      "get": {
        "description": "같은 연령대 또래 대비 비교 — 또래 평균·상위%·메달(참고용 추정).",
        "operationId": "get_peer_comparison_v1_growth_peers_get",
        "parameters": [
          {
            "description": "User identification key",
            "in": "header",
            "name": "x-user-key",
            "required": true,
            "schema": {
              "description": "User identification key",
              "title": "X-User-Key",
              "type": "string"
            }
          },
          {
            "description": "Optional active child profile id",
            "in": "header",
            "name": "x-profile-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional active child profile id",
              "title": "X-Profile-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Peer Comparison",
        "tags": [
          "Growth"
        ]
      }
    },
...
    "/v1/profiles": {
      "get": {
        "operationId": "list_profiles_v1_profiles_get",
        "parameters": [
          {
            "description": "User identification key",
            "in": "header",
            "name": "x-user-key",
            "required": true,
            "schema": {
              "description": "User identification key",
              "title": "X-User-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Profiles",
        "tags": [
          "Profiles"
        ]
      },
      "post": {
        "operationId": "create_profile_v1_profiles_post",
        "parameters": [
          {
            "description": "User identification key",
            "in": "header",
            "name": "x-user-key",
            "required": true,
            "schema": {
              "description": "User identification key",
              "title": "X-User-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Profile",
        "tags": [
          "Profiles"
        ]
      }
    },
    "/v1/profiles/{profile_id}": {
      "delete": {
        "operationId": "delete_profile_v1_profiles__profile_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "title": "Profile Id",
              "type": "string"
            }
          },
          {
            "description": "User identification key",
            "in": "header",
            "name": "x-user-key",
            "required": true,
            "schema": {
              "description": "User identification key",
              "title": "X-User-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Profile",
        "tags": [
          "Profiles"
        ]
      },
      "patch": {
        "operationId": "update_profile_v1_profiles__profile_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "title": "Profile Id",
              "type": "string"
            }
          },
          {
            "description": "User identification key",
            "in": "header",
            "name": "x-user-key",
            "required": true,
            "schema": {
              "description": "User identification key",
              "title": "X-User-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Profile",
        "tags": [
          "Profiles"
        ]
      }
    },
...
    "/v1/streak/info": {
      "get": {
        "description": "스트릭 정보 조회\n\n- current_streak: 현재 연속 읽기 일수\n- longest_streak: 최장 연속 일수\n- total_days: 총 읽은 일수\n- read_today: 오늘 읽었는지 여부",
        "operationId": "get_streak_info_v1_streak_info_get",
        "parameters": [
          {
            "description": "User identification key",
            "in": "header",
            "name": "x-user-key",
            "required": true,
            "schema": {
              "description": "User identification key",
              "title": "X-User-Key",
              "type": "string"
            }
          },
          {
            "description": "Optional active child profile id",
            "in": "header",
            "name": "x-profile-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional active child profile id",
              "title": "X-Profile-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StreakInfoResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Streak Info",
        "tags": [
          "Streak"
        ]
      }
    },
    "/v1/streak/read": {
      "post": {
        "description": "읽기 기록\n\n- 책을 읽었을 때 호출\n- 오늘 처음 읽는 경우 스트릭 증가\n- 마일스톤 달성 시 알림",
        "operationId": "record_reading_v1_streak_read_post",
        "parameters": [
          {
            "description": "User identification key",
            "in": "header",
            "name": "x-user-key",
            "required": true,
            "schema": {
              "description": "User identification key",
              "title": "X-User-Key",
              "type": "string"
            }
          },
          {
            "description": "Optional active child profile id",
            "in": "header",
            "name": "x-profile-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional active child profile id",
              "title": "X-Profile-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReadingLogRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadingResultResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Record Reading",
        "tags": [
          "Streak"
        ]
      }
    },
...
  },
...
}